Add information for Onyx environment
I don't think this information warrants being put in the readme, but Composer also needs to be installed locally by the user on Onyx following the instructions on https://getcomposer.org/doc/00-intro.md#locally. Once installed it needs either its location added to PATH and the file composer.phar renamed to simply composer, or an alias set up so that "composer" calls composer.phar from wherever it was installed. Whichever change is chosen, this should also be added to .bash_profile.